fix: correct DWORD/%x format specifier mismatch in NONCONTINUABLEDIAG
DWORD is unsigned long, not unsigned int, so passing ExceptionFlags/
ExceptionCode fields directly to a %x format specifier is UB and fails
-Werror=format on both MSVC/clang-cl (Windows Run Clang Tidy) and
GCC/MinGW builds. Cast to unsigned int explicitly.

+++ src/windows-emulator/exception_dispatch.cpp
@@ -252,8 +252,10 @@ namespace sogen
                 fprintf(stderr,
                         "[NONCONTINUABLEDIAG] wow64_record.ExceptionFlags=0x%x wow64_record.ExceptionCode=0x%x record_addr=0x%llx "
                         "readback_ok=%d readback.ExceptionFlags=0x%x readback.ExceptionCode=0x%x\n",
-                        wow64_record.ExceptionFlags, wow64_record.ExceptionCode, static_cast<unsigned long long>(record_addr),
-                        readback.has_value() ? 1 : 0, readback ? readback->ExceptionFlags : 0, readback ? readback->ExceptionCode : 0);
+                        static_cast<unsigned int>(wow64_record.ExceptionFlags), static_cast<unsigned int>(wow64_record.ExceptionCode),
+                        static_cast<unsigned long long>(record_addr), readback.has_value() ? 1 : 0,
+                        static_cast<unsigned int>(readback ? readback->ExceptionFlags : 0),
+                        static_cast<unsigned int>(readback ? readback->ExceptionCode : 0));
                 fflush(stderr);
             }
